Validações nos cadastros, Janela de login arrumada, alterações de variaveis (de int para long), porque o valor sempre passava e dava erro.

+++ PetShop/src/petshop/classes/ItemVenda.java
@@ -6,7 +6,7 @@ package petshop.classes;
  */
 public class ItemVenda {
 	
-	private int codigo;
+	private long codigo;
 	private String nome;
 	private double precoVenda;
 	
@@ -16,7 +16,7 @@ public class ItemVenda {
 	 * @param nome
 	 * @param precoVenda
 	 */
-	public ItemVenda(int codigo, String nome, double precoVenda) {
+	public ItemVenda(long codigo, String nome, double precoVenda) {
 		super();
 		this.codigo = codigo;
 		this.nome = nome;
@@ -29,7 +29,7 @@ public class ItemVenda {
 	/**
 	 * @return
 	 */
-	public int getCodigo() {
+	public long getCodigo() {
 		return codigo;
 	}
 	/**
